64 bit fix
git-svn-id: http://shedskin.googlecode.com/svn/trunk@262 acccb5f4-fd3f-0410-a51f-a7a3fbc0418d

+++ lib/builtin.hpp
@@ -1117,7 +1117,7 @@ template<class K, class V> dict<K, V>::dict(PyObject *p) {
     
     this->__class__ = cl_dict;
     PyObject *key, *value;
-    int pos = 0;
+    Py_ssize_t pos = 0;
     while(PyDict_Next(p, &pos, &key, &value)) 
         __setitem__(__to_ss<K>(key), __to_ss<V>(value));
 }
